refactor: Reorganize cascade interactions
