+++ include/Achilles/Interactions.hh
@@ -1,27 +1,8 @@
 #pragma once
 
-#include <array>
-#include <map>
-#include <memory>
-#include <unordered_map>
 #include <vector>
 
 #include "Achilles/Factory.hh"
-#include "Achilles/Interpolation.hh"
-#include "Achilles/MesonBaryonAmplitudes.hh"
-#include "Achilles/OsetAbsorption.hh"
-#include "Achilles/Particle.hh"
-#include "Achilles/ParticleInfo.hh"
-#include "Achilles/ThreeVector.hh"
-
-#pragma GCC diagnostic push
-#pragma GCC diagnostic ignored "-Wold-style-cast"
-#ifndef __clang__
-#pragma GCC diagnostic ignored "-Wuseless-cast"
-#endif
-#pragma GCC diagnostic ignored "-Wnull-dereference"
-#include "highfive/H5Group.hpp"
-#pragma GCC diagnostic pop
 
 #pragma GCC diagnostic push
 #pragma GCC diagnostic ignored "-Wshadow"
@@ -31,8 +12,8 @@
 namespace achilles {
 
 class Event;
-class FourVector;
-class Potential;
+class PID;
+class Particle;
 class Random;
 
 struct pid_compare {
@@ -107,336 +88,6 @@ class Interaction {
 template <typename Derived>
 using RegistrableInteraction = Registrable<Interaction, Derived, const YAML::Node &>;
 using InteractionFactory = Factory<Interaction, const YAML::Node &>;
-
-class NasaInteraction : public Interaction, RegistrableInteraction<NasaInteraction> {
-  public:
-    NasaInteraction() = default;
-    NasaInteraction(const YAML::Node &){};
-
-    static std::unique_ptr<Interaction> Construct(const YAML::Node &data) {
-        return std::make_unique<NasaInteraction>(data);
-    }
-
-    /// Returns the name of the class, used in the InteractionFactory
-    ///@return std::string: The name of the class
-    std::string GetName() const override { return NasaInteraction::Name(); }
-    static std::string Name() { return "NasaInteraction"; }
-
-    // These functions are defined in the base class
-    std::vector<std::pair<PID, PID>> InitialStates() const override {
-        return {{PID::proton(), PID::proton()},
-                {PID::neutron(), PID::proton()},
-                {PID::neutron(), PID::neutron()}};
-    }
-    InteractionResults CrossSection(Event &, size_t, size_t) const override;
-    std::vector<Particle> GenerateMomentum(const Particle &part1, const Particle &part2,
-                                           const std::vector<PID> &out_pids,
-                                           Random &) const override;
-
-  private:
-    ThreeVector MakeMomentum(bool, double, const std::vector<double> &) const;
-};
-
-/// Class for implementing an interaction model based on the Geant4 cross-section data. This
-/// interaction model contains information about the angular distribution of pp, pn, and nn
-/// interactions that occur during the intranuclear cascade.
-class GeantInteraction : public Interaction, RegistrableInteraction<GeantInteraction> {
-  public:
-    ///@name Constructors and Destructors
-    ///@{
-
-    /// Initialize GeantInteractions class. This loads data from an input file
-    ///@param filename: The location of the Geant4 hdf5 data file
-    GeantInteraction(const YAML::Node &);
-    GeantInteraction(const GeantInteraction &) = default;
-    GeantInteraction(GeantInteraction &&) = default;
-    GeantInteraction &operator=(const GeantInteraction &) = default;
-    GeantInteraction &operator=(GeantInteraction &&) = default;
-
-    /// Generate a GeantInteractions object. This is used in the InteractionFactory.
-    ///@param data: The location of the data file to load containing the Geant4 cross-sections
-    static std::unique_ptr<Interaction> Construct(const YAML::Node &data) {
-        return std::make_unique<GeantInteraction>(data);
-    }
-
-    /// Default Destructor
-    ~GeantInteraction() override = default;
-    ///@}
-
-    /// Returns the name of the class, used in the InteractionFactory
-    ///@return std::string: The name of the class
-    std::string GetName() const override { return GeantInteraction::Name(); }
-    static std::string Name() { return "GeantInteraction"; }
-
-    // These functions are defined in the base class
-    std::vector<std::pair<PID, PID>> InitialStates() const override {
-        return {{PID::proton(), PID::proton()},
-                {PID::neutron(), PID::proton()},
-                {PID::neutron(), PID::neutron()}};
-    }
-    InteractionResults CrossSection(Event &, size_t, size_t) const override;
-    std::vector<Particle> GenerateMomentum(const Particle &part1, const Particle &part2,
-                                           const std::vector<PID> &out_pids,
-                                           Random &) const override;
-
-  private:
-    // Functions
-    double CrossSectionAngle(bool, const double &, const double &) const;
-    void LoadData(bool, const HighFive::Group &);
-    ThreeVector MakeMomentum(bool, double, const std::vector<double> &) const;
-
-    // Variables
-    std::vector<double> m_theta, m_cdf;
-    std::vector<double> m_pcmPP, m_xsecPP;
-    std::vector<double> m_pcmNP, m_xsecNP;
-    Interp1D m_crossSectionPP, m_crossSectionNP;
-    Interp2D m_thetaDistPP, m_thetaDistNP;
-};
-
-class MesonBaryonInteraction : public Interaction, RegistrableInteraction<MesonBaryonInteraction> {
-  public:
-    ///@name Constructors and Destructors
-    ///@{
-
-    /// Initialize MesonBaryonInteraction class. This loads data from an input file
-    MesonBaryonInteraction() = default;
-    MesonBaryonInteraction(const YAML::Node &){};
-
-    /// Generate a object. This is used in the InteractionFactory.
-    static std::unique_ptr<Interaction> Construct(const YAML::Node &data) {
-        return std::make_unique<MesonBaryonInteraction>(data);
-    }
-
-    /// Default Destructor
-    ~MesonBaryonInteraction() override = default;
-    ///@}
-
-    /// Returns the name of the class, used in the InteractionFactory
-    ///@return std::string: The name of the class
-    std::string GetName() const override { return MesonBaryonInteraction::Name(); }
-    static std::string Name() { return "MesonBaryonInteraction"; }
-
-    // List off all InitialStates
-    std::vector<std::pair<PID, PID>> InitialStates() const override {
-        std::vector<std::pair<PID, PID>> states;
-
-        int nChan = m_amps.NChargeChannels();
-
-        for(int i = 0; i < nChan; i++) {
-            int pidm = m_amps.MesonPID_Cchan(i);
-            int pidb = m_amps.BaryonPID_Cchan(i);
-            states.push_back({PID(pidm), PID(pidb)});
-        }
-        return states;
-    }
-
-    // Cross sections and pairs of PIDs for all possible final states
-    InteractionResults CrossSection(Event &, size_t, size_t) const override;
-
-    std::vector<Particle> GenerateMomentum(const Particle &part1, const Particle &part2,
-                                           const std::vector<PID> &out_pids,
-                                           Random &) const override;
-
-  private:
-    // Variables
-    MBAmplitudes m_amps; // Contains all the mesonBaryon information
-};
-
-/// Class for implementing an interaction model based on the Geant4 cross-section data. This
-/// interaction model contains information about the angular distribution of pp, pn, and nn
-/// interactions that occur during the intranuclear cascade.
-class ConstantInteraction : public Interaction, RegistrableInteraction<ConstantInteraction> {
-  public:
-    ///@name Constructors and Destructors
-    ///@{
-
-    /// Initialize ConstantInteractions class for testing purposes.
-    ConstantInteraction() = default;
-    /// Initialize ConstantInteractions class. This loads data from an input file
-    ///@param node: The constant interactions to load
-    ConstantInteraction(const YAML::Node &);
-    ConstantInteraction(const ConstantInteraction &) = default;
-    ConstantInteraction(ConstantInteraction &&) = default;
-    ConstantInteraction &operator=(const ConstantInteraction &) = default;
-    ConstantInteraction &operator=(ConstantInteraction &&) = default;
-
-    /// Generate a ConstantInteractions object. This is used in the InteractionFactory.
-    ///@param data: The location of the data file to load containing the Geant4 cross-sections
-    static std::unique_ptr<Interaction> Construct(const YAML::Node &data) {
-        return std::make_unique<ConstantInteraction>(data);
-    }
-
-    /// Default Destructor
-    ~ConstantInteraction() override = default;
-    ///@}
-
-    /// Returns the name of the class, used in the InteractionFactory
-    ///@return std::string: The name of the class
-    std::string GetName() const override { return ConstantInteraction::Name(); }
-    static std::string Name() { return "ConstantInteraction"; }
-
-    // These functions are defined in the base class
-    std::vector<std::pair<PID, PID>> InitialStates() const override { return m_states; }
-    InteractionResults CrossSection(Event &, size_t, size_t) const override;
-    std::vector<Particle> GenerateMomentum(const Particle &part1, const Particle &part2,
-                                           const std::vector<PID> &out_pids,
-                                           Random &) const override;
-
-    // These functions are for testing only
-    void AddInteraction(const std::pair<PID, PID> &state, const InteractionResults &results) {
-        if(m_interactions.find(state) != m_interactions.end()) {
-            auto msg =
-                fmt::format("Initial state: [{}, {}] already exists", state.first, state.second);
-            throw std::runtime_error("Initial state already exists");
-        }
-        m_interactions[state] = results;
-        m_states.push_back(state);
-    }
-
-  private:
-    ThreeVector MakeMomentum(bool, double, const std::vector<double> &) const;
-
-    // Variables
-    std::map<std::pair<PID, PID>, InteractionResults, pid_compare> m_interactions;
-    std::vector<std::pair<PID, PID>> m_states;
-};
-
-class PionAbsorption : public Interaction {
-  public:
-    /// Initialize PionAbsorption Helper class
-    PionAbsorption() = default;
-    PionAbsorption(const YAML::Node &);
-
-    // List off all InitialStates
-    std::vector<std::pair<PID, PID>> InitialStates() const override {
-        return {{PID::pion0(), PID::proton()},  {PID::pionp(), PID::proton()},
-                {-PID::pionp(), PID::proton()}, {PID::pion0(), PID::neutron()},
-                {PID::pionp(), PID::neutron()}, {-PID::pionp(), PID::neutron()}};
-    }
-
-    InteractionResults CrossSection(Event &, size_t, size_t) const override;
-
-  protected:
-    virtual bool AllowedAbsorption(Event &, size_t, size_t) const = 0;
-    mutable Particle absorption_partner;
-
-    OsetAbsCrossSection Oset_abs; // Contains Oset absorption cross sections
-
-    struct InteractionStates {
-        PID absorption_partner;
-        std::vector<PID> outgoing;
-    };
-
-    using InteractionMap = std::map<std::pair<PID, PID>, std::vector<InteractionStates>>;
-    InteractionMap states;
-};
-
-class PionAbsorptionOneStep : public PionAbsorption, RegistrableInteraction<PionAbsorptionOneStep> {
-  public:
-    ///@name Constructors and Destructors
-    ///@{
-
-    /// Initialize PionAbsorptionOneStep class
-    PionAbsorptionOneStep() = default;
-    PionAbsorptionOneStep(const YAML::Node &){};
-
-    /// Generate a object. This is used in the InteractionFactory.
-    static std::unique_ptr<Interaction> Construct(const YAML::Node &data) {
-        return std::make_unique<PionAbsorptionOneStep>(data);
-    }
-
-    /// Default Destructor
-    ~PionAbsorptionOneStep() override = default;
-    ///@}
-
-    /// Returns the name of the class, used in the InteractionFactory
-    ///@return std::string: The name of the class
-    std::string GetName() const override { return PionAbsorptionOneStep::Name(); }
-    static std::string Name() { return "PionAbsorptionOneStep"; }
-
-    std::vector<Particle> GenerateMomentum(const Particle &part1, const Particle &part2,
-                                           const std::vector<PID> &out_pids,
-                                           Random &) const override;
-
-  private:
-    std::pair<size_t, size_t> FindClosest(Event &, size_t, size_t) const;
-    bool AllowedAbsorption(Event &, size_t, size_t) const override;
-    mutable std::vector<std::pair<size_t, std::vector<PID>>> m_states;
-};
-
-class PionAbsorptionTwoStep : public PionAbsorption, RegistrableInteraction<PionAbsorptionTwoStep> {
-  public:
-    ///@name Constructors and Destructors
-    ///@{
-
-    /// Initialize PionAbsorptionTwoStep class
-    PionAbsorptionTwoStep() = default;
-    PionAbsorptionTwoStep(const YAML::Node &){};
-
-    /// Generate a object. This is used in the InteractionFactory.
-    static std::unique_ptr<Interaction> Construct(const YAML::Node &data) {
-        return std::make_unique<PionAbsorptionTwoStep>(data);
-    }
-
-    /// Default Destructor
-    ~PionAbsorptionTwoStep() override = default;
-    ///@}
-
-    /// Returns the name of the class, used in the InteractionFactory
-    ///@return std::string: The name of the class
-    std::string GetName() const override { return PionAbsorptionTwoStep::Name(); }
-    static std::string Name() { return "PionAbsorptionTwoStep"; }
-
-    std::vector<Particle> GenerateMomentum(const Particle &part1, const Particle &part2,
-                                           const std::vector<PID> &out_pids,
-                                           Random &) const override;
-
-  private:
-    bool CheckHistory(Event &) const;
-    bool AllowedAbsorption(Event &, size_t, size_t) const override;
-};
-
-class PionInteraction : public Interaction, RegistrableInteraction<PionInteraction> {
-  public:
-    ///@name Constructors and Destructors
-    ///@{
-
-    /// Initialize PionAbsorptionTwoStep class
-    PionInteraction() = default;
-    PionInteraction(const YAML::Node &);
-
-    /// Generate a object. This is used in the InteractionFactory.
-    static std::unique_ptr<Interaction> Construct(const YAML::Node &data) {
-        return std::make_unique<PionInteraction>(data);
-    }
-
-    /// Default Destructor
-    ~PionInteraction() override = default;
-    ///@}
-
-    /// Returns the name of the class, used in the InteractionFactory
-    ///@return std::string: The name of the class
-    std::string GetName() const override { return PionInteraction::Name(); }
-    static std::string Name() { return "PionInteraction"; }
-
-    // List off all InitialStates
-    std::vector<std::pair<PID, PID>> InitialStates() const override {
-        return {{PID::pion0(), PID::proton()},  {PID::pionp(), PID::proton()},
-                {-PID::pionp(), PID::proton()}, {PID::pion0(), PID::neutron()},
-                {PID::pionp(), PID::neutron()}, {-PID::pionp(), PID::neutron()}};
-    }
-
-    InteractionResults CrossSection(Event &, size_t, size_t) const override;
-
-    std::vector<Particle> GenerateMomentum(const Particle &part1, const Particle &part2,
-                                           const std::vector<PID> &out_pids,
-                                           Random &) const override;
-
-  private:
-    std::unique_ptr<Interaction> hard_scatter;
-    // TODO: Determine if this can remain an Interaction ptr or needs to be an Absorption ptr
-    std::unique_ptr<Interaction> absorption;
-};
 
 } // namespace achilles
